% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.r
\name{contract}
\alias{contract}
\title{Get the Contracted Network}
\usage{
contract(network)
}
\arguments{
\item{network}{A network data frame in the dodgr class (refer to the "dodgr"
package).}
}
\value{
The output is a contracted network.
}
\description{
This function introduces dodgr_contract_graph functions to get the major
contracted network with "component = 1".
}
\examples{
location <- c(osmdata::getbb("oxford uk"))
wholenet <- getnet("oxford uk")
mainnet <- mainnet(wholenet)
cntrnet <- contract(mainnet)
shownetwork(vec2bbox(location), wholenet, cntrnet)

}
\keyword{contract}
